refactor image and video draft handling
