+++ front-end/src/components/EditorContainer.jsx
@@ -1,44 +1,27 @@
 import { useCallback, useEffect, useState } from 'react'
-import { createCreation, updateCreation } from '../services/creationsApi.js'
-import { buildImageCreationPayload, buildVideoCreationPayload, defaultCreationTitle } from '../utils/buildCreationPayload.js'
 import { resolveDraftMediaIds } from '../utils/draftMediaIds.js'
-import { getOrCreateOwnerKey } from '../utils/ownerKey.js'
+import { getImageDraftRestoreState, getVideoDraftRestoreState } from '../utils/draftRestoreState.js'
+import { convertBackendVideoResultToLocalMedia } from '../services/backendGifService'
+import { getBackendBaseUrl } from '../services/backendMediaClient'
 import MediaEntry from './MediaEntry'
-import ImageEditor from './image/ImageEditor'
-import FilterMain from './FilterMain'
-import PresetFilters from './image/PresetFilters'
-import PresetSizes from './image/PresetSizes'
-import AddText from './image/AddText'
-import ColorFilters from './image/ColorFilters'
-import GifEditor from './gif/GifEditor'
+import CameraCapture from './CameraCapture'
+import PhotoPreview from './PhotoPreview'
+import EditorStatus from './EditorStatus.jsx'
+import ImageEditorFlow from './image/ImageEditorFlow'
+import GifEditorFlow from './gif/GifEditorFlow'
 import useMediaSelection, { MEDIA_SELECTION_CODES } from '../hooks/useMediaSelection'
 import useGifEditingSession from '../hooks/useGifEditingSession'
 import useImageEditingSession from '../hooks/useImageEditingSession'
-import { convertBackendVideoResultToLocalMedia } from '../services/backendGifService'
-import { getBackendBaseUrl } from '../services/backendMediaClient'
-import CameraCapture from './CameraCapture'
-import PhotoPreview from './PhotoPreview'
-import VideoPresetFilters from './gif/VideoPresetFilters'
-import GifFilterMain from './gif/GifFilterMain'
-import GifSpeedControls from './gif/GifSpeedControls'
-import GifTrimEditor from './gif/GifTrimEditor'
-import GifResizePresets from './gif/GifResizePresets'
-import GifTextOverlayEditor from './gif/GifTextOverlayEditor'
-import EditorStatus from './EditorStatus.jsx'
 
 const SCREENS = {
   EDITOR: 'editor',
-  FILTERS_MAIN: 'filters-main',
-  IMAGE_PRESET_FILTERS: 'image-preset-filters',
-  ADD_TEXT: 'text',
-  COLOR_FILTERS: 'color',
-  PRESET_SIZES: 'preset-sizes',
   CAMERA: 'camera',
   CAMERA_PREVIEW: 'camera-preview',
 }
 
 const FILE_TOO_LARGE_MESSAGE = 'File is too large (max 50 MB).'
-const UNSUPPORTED_IMAGE_MESSAGE = 'GIF and HEIC/HEIF files are not supported for image editing yet. Please upload JPG or PNG.'
+const UNSUPPORTED_IMAGE_MESSAGE =
+  'GIF and HEIC/HEIF files are not supported for image editing yet. Please upload JPG or PNG.'
 
 function EditorContainer({ onDraftSaved, onSelectCreation }) {
   const {
@@ -61,45 +44,19 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
 
   const gifSession = useGifEditingSession()
 
-  const {
-    letterboxColor,
-    setLetterboxColor,
-    isExporting,
-    isResettingCrop,
-    exportError,
-    lastCropBoxPx,
-    hasCropHistory,
-    effectiveImageSrc,
-    selectedPreset,
-    effectiveBackendMediaId,
-    colorAdjustments,
-    selectedImageFilterPreset,
-    resetImageEditingSessionState,
-    presetFilterPreviewSrc,
-    isLoadingPresetFilterPreview,
-    presetFilterError,
-    colorFilterPreviewSrc,
-    isLoadingColorFilterPreview,
-    colorFilterError,
-    clearCropSession,
-    handleSizeSelect,
-    handleCropApply,
-    resetCropToOriginal,
-    handleAddTextApply,
-    handleExport,
-    selectImageFilterPreset,
-    applyImagePresetFilter,
-    updateColorAdjustments,
-    applyColorAdjustments,
-    restoreImageSession,
-    editBaseMediaId,
-  } = useImageEditingSession({
+  const imageSession = useImageEditingSession({
     mediaType,
     backendImageResult,
     previewUrl,
     sourceUrl,
     applyTransformedImage,
   })
+
+  const {
+    resetImageEditingSessionState,
+    clearCropSession,
+    restoreImageSession,
+  } = imageSession
 
   const [screen, setScreen] = useState(SCREENS.EDITOR)
   const [activeDraftId, setActiveDraftId] = useState(null)
@@ -108,9 +65,6 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
   const [videoDraftSourceMediaId, setVideoDraftSourceMediaId] = useState(null)
   const [appliedTextOverlay, setAppliedTextOverlay] = useState(null)
   const [preTextWorkingMediaId, setPreTextWorkingMediaId] = useState(null)
-  const [saveForLaterError, setSaveForLaterError] = useState(null)
-  const [saveForLaterMessage, setSaveForLaterMessage] = useState(null)
-  const [isSavingDraft, setIsSavingDraft] = useState(false)
   const [fileTooLargeMessage, setFileTooLargeMessage] = useState(null)
   const [unsupportedImageMessage, setUnsupportedImageMessage] = useState(null)
   const [lastRejectedUploadType, setLastRejectedUploadType] = useState(null)
@@ -123,14 +77,17 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
   const [unsupportedVideo, setUnsupportedVideo] = useState(null)
 
   const effectiveImageDraftSourceMediaId =
-    imageDraftSourceMediaId ||
-    (mediaType === 'image' ? backendImageResult?.id : null)
+    imageDraftSourceMediaId || (mediaType === 'image' ? backendImageResult?.id : null)
 
   const effectiveVideoDraftSourceMediaId =
-    videoDraftSourceMediaId ||
-    (mediaType === 'video' ? backendVideoResult?.id : null)
+    videoDraftSourceMediaId || (mediaType === 'video' ? backendVideoResult?.id : null)
 
   const shouldShowDraftLoadCancel = isDraftLoading && showDraftLoadCancel
+
+  const handleActiveDraftSaved = useCallback((id, title) => {
+    if (id) setActiveDraftId(id)
+    if (title) setActiveDraftTitle(title)
+  }, [])
 
   const handleImageSelect = async (file) => {
     if (!file) return
@@ -163,8 +120,6 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
     setActiveDraftTitle(null)
     setImageDraftSourceMediaId(null)
     setVideoDraftSourceMediaId(null)
-    setSaveForLaterError(null)
-    setSaveForLaterMessage(null)
     setAppliedTextOverlay(null)
     setPreTextWorkingMediaId(null)
 
@@ -174,7 +129,14 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
   }
 
   const handleVideoSelect = useCallback(
-    async (file, { preserveSelectedSpeed = false, preserveOriginalVideo = false, preserveDraftIdentity = false } = {}) => {
+    async (
+      file,
+      {
+        preserveSelectedSpeed = false,
+        preserveOriginalVideo = false,
+        preserveDraftIdentity = false,
+      } = {},
+    ) => {
       if (!file) return
 
       setLastRejectedUploadType('video')
@@ -212,8 +174,6 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
           setImageDraftSourceMediaId(null)
           setVideoDraftSourceMediaId(null)
         }
-        setSaveForLaterError(null)
-        setSaveForLaterMessage(null)
 
         if (pendingVideoDraftPayload) {
           gifSession.restoreGifSession(pendingVideoDraftPayload)
@@ -246,150 +206,169 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
     setActiveDraftTitle(null)
     setImageDraftSourceMediaId(null)
     setVideoDraftSourceMediaId(null)
-    setSaveForLaterError(null)
-    setSaveForLaterMessage(null)
     setAppliedTextOverlay(null)
     setPreTextWorkingMediaId(null)
     setShowDraftLoadCancel(false)
     setScreen(SCREENS.EDITOR)
   }
 
-  const handleLoadDraft = useCallback(async (creation) => {
-    const payload = creation?.editorPayload
-    if (!payload) return
+  const handleLoadDraft = useCallback(
+    async (creation) => {
+      const payload = creation?.editorPayload
+      if (!payload) return
 
-    setDraftLoadError(null)
-    setShowDraftLoadCancel(false)
+      setDraftLoadError(null)
+      setShowDraftLoadCancel(false)
 
-    const draftId = String(creation._id ?? creation.id)
-    const nextDraftTitle = typeof creation?.title === 'string' && creation.title.trim()
-      ? creation.title.trim()
-      : null
-    const { sourceMediaId, workingMediaId, resumeMediaId } = resolveDraftMediaIds(payload)
+      const draftId = String(creation._id ?? creation.id)
+      const nextDraftTitle =
+        typeof creation?.title === 'string' && creation.title.trim() ? creation.title.trim() : null
+      const { sourceMediaId, workingMediaId, resumeMediaId } = resolveDraftMediaIds(payload)
 
-    if (payload.kind === 'image') {
-      setIsDraftLoading(true)
-      resetSelection()
-      resetImageEditingSessionState()
-      gifSession.resetGifSession()
-      setPendingVideoDraftPayload(null)
-      setActiveDraftId(draftId)
-      setActiveDraftTitle(nextDraftTitle)
-      setImageDraftSourceMediaId(sourceMediaId)
-      setVideoDraftSourceMediaId(null)
-      setSaveForLaterError(null)
-      setSaveForLaterMessage(null)
+      if (payload.kind === 'image') {
+        const imageRestore = getImageDraftRestoreState(payload)
 
-      const ok = await restoreImageSession(payload)
-      setIsDraftLoading(false)
+        setIsDraftLoading(true)
+        resetSelection()
+        resetImageEditingSessionState()
+        gifSession.resetGifSession()
+        setPendingVideoDraftPayload(null)
+        setActiveDraftId(draftId)
+        setActiveDraftTitle(nextDraftTitle)
+        setImageDraftSourceMediaId(sourceMediaId)
+        setVideoDraftSourceMediaId(null)
 
-      if (ok) {
-        setAppliedTextOverlay(payload.textOverlay ?? null)
-        setPreTextWorkingMediaId(payload.preTextWorkingMediaId ?? null)
-        setScreen(SCREENS.EDITOR)
-      } else {
-        setDraftLoadError('Could not restore draft image. The file may have expired.')
-        setActiveDraftId(null)
-        setActiveDraftTitle(null)
-        setImageDraftSourceMediaId(null)
-        setAppliedTextOverlay(null)
-        setPreTextWorkingMediaId(null)
-      }
-      return
-    }
+        const normalizedPayload = {
+          ...payload,
+          lastCropBoxPx: imageRestore.crop,
+          colorAdjustments: imageRestore.colorAdjustments,
+          selectedImageFilterPreset: imageRestore.presetFilter,
+          textOverlay: imageRestore.textOverlay,
+          selectedPreset: imageRestore.resize?.preset ?? payload?.selectedPreset ?? null,
+          letterboxColor: imageRestore.resize?.letterboxColor ?? payload?.letterboxColor ?? 'transparent',
+        }
 
-    if (payload.kind === 'video') {
-      resetSelection()
-      resetImageEditingSessionState()
-      gifSession.resetGifSession()
-      setActiveDraftId(draftId)
-      setActiveDraftTitle(nextDraftTitle)
-      setImageDraftSourceMediaId(null)
-      setVideoDraftSourceMediaId(sourceMediaId)
-      setSaveForLaterError(null)
-      setSaveForLaterMessage(null)
-      setPendingVideoDraftPayload(null)
+        const ok = await restoreImageSession(normalizedPayload)
+        setIsDraftLoading(false)
 
-      if (!resumeMediaId) {
-        setPendingVideoDraftPayload(payload)
+        if (ok) {
+          setAppliedTextOverlay(imageRestore.textOverlay ?? null)
+          setPreTextWorkingMediaId(payload.preTextWorkingMediaId ?? null)
+          setScreen(SCREENS.EDITOR)
+        } else {
+          setDraftLoadError('Could not restore draft image. The file may have expired.')
+          setActiveDraftId(null)
+          setActiveDraftTitle(null)
+          setImageDraftSourceMediaId(null)
+          setAppliedTextOverlay(null)
+          setPreTextWorkingMediaId(null)
+        }
         return
       }
 
-      setIsDraftLoading(true)
+      if (payload.kind === 'video') {
+        const videoRestore = getVideoDraftRestoreState(payload)
+        const normalizedPayload = {
+          ...payload,
+          trimRange: videoRestore.trimRange,
+          resizePreset: videoRestore.resizePreset,
+          resizeBorderColor: videoRestore.resizeBorderColor,
+          selectedSpeedPlaybackRate: videoRestore.selectedSpeedPlaybackRate,
+          textOverlaySettings: videoRestore.textOverlaySettings,
+          selectedFilterPreset: videoRestore.selectedFilterPreset,
+        }
 
-      try {
-        setPendingVideoDraftPayload(payload)
+        resetSelection()
+        resetImageEditingSessionState()
+        gifSession.resetGifSession()
+        setActiveDraftId(draftId)
+        setActiveDraftTitle(nextDraftTitle)
+        setImageDraftSourceMediaId(null)
+        setVideoDraftSourceMediaId(sourceMediaId)
+        setPendingVideoDraftPayload(null)
 
-        const candidateIds = [...new Set([workingMediaId, sourceMediaId].filter(Boolean))]
-        let restoredFile = null
+        if (!resumeMediaId) {
+          setPendingVideoDraftPayload(normalizedPayload)
+          return
+        }
 
-        for (const mediaId of candidateIds) {
-          try {
-            const restoredResult = {
-              id: mediaId,
-              url: `${getBackendBaseUrl()}/api/media/${mediaId}`,
-              mimeType: 'video/mp4',
+        setIsDraftLoading(true)
+
+        try {
+          setPendingVideoDraftPayload(normalizedPayload)
+
+          const candidateIds = [...new Set([workingMediaId, sourceMediaId].filter(Boolean))]
+          let restoredFile = null
+
+          for (const mediaId of candidateIds) {
+            try {
+              const restoredResult = {
+                id: mediaId,
+                url: `${getBackendBaseUrl()}/api/media/${mediaId}`,
+                mimeType: 'video/mp4',
+              }
+
+              const converted = await convertBackendVideoResultToLocalMedia(restoredResult, {
+                fallbackFileName: 'draft-video.mp4',
+                fallbackMimeType: 'video/mp4',
+                fetchErrorMessage: 'Failed to load draft video.',
+              })
+
+              restoredFile = converted.file
+              break
+            } catch {
+              // Try next candidate id.
             }
-
-            const converted = await convertBackendVideoResultToLocalMedia(restoredResult, {
-              fallbackFileName: 'draft-video.mp4',
-              fallbackMimeType: 'video/mp4',
-              fetchErrorMessage: 'Failed to load draft video.',
-            })
-
-            restoredFile = converted.file
-            break
-          } catch {
-            // Try next candidate id.
           }
-        }
 
-        if (!restoredFile) {
-          throw new Error('Draft video could not be restored.')
-        }
+          if (!restoredFile) {
+            throw new Error('Draft video could not be restored.')
+          }
 
-        const selection = await handleVideoSelect(restoredFile, {
-          preserveSelectedSpeed: true,
-          preserveOriginalVideo: false,
-          preserveDraftIdentity: true,
-        })
+          const selection = await handleVideoSelect(restoredFile, {
+            preserveSelectedSpeed: true,
+            preserveOriginalVideo: false,
+            preserveDraftIdentity: true,
+          })
 
-        if (selection?.code !== MEDIA_SELECTION_CODES.OK) {
-          throw new Error('Draft video could not be restored.')
-        }
+          if (selection?.code !== MEDIA_SELECTION_CODES.OK) {
+            throw new Error('Draft video could not be restored.')
+          }
 
-        if (sourceMediaId && workingMediaId && sourceMediaId !== workingMediaId) {
-          try {
-            const sourceResult = {
-              id: sourceMediaId,
-              url: `${getBackendBaseUrl()}/api/media/${sourceMediaId}`,
-              mimeType: 'video/mp4',
+          if (sourceMediaId && workingMediaId && sourceMediaId !== workingMediaId) {
+            try {
+              const sourceResult = {
+                id: sourceMediaId,
+                url: `${getBackendBaseUrl()}/api/media/${sourceMediaId}`,
+                mimeType: 'video/mp4',
+              }
+              const { file: sourceFile } = await convertBackendVideoResultToLocalMedia(sourceResult, {
+                fallbackFileName: 'source-video.mp4',
+                fallbackMimeType: 'video/mp4',
+                fetchErrorMessage: 'Failed to load source video.',
+              })
+              setOriginalVideoFile(sourceFile)
+            } catch {
+              // Non-critical. originalVideoFile stays as the working video.
             }
-            const { file: sourceFile } = await convertBackendVideoResultToLocalMedia(sourceResult, {
-              fallbackFileName: 'source-video.mp4',
-              fallbackMimeType: 'video/mp4',
-              fetchErrorMessage: 'Failed to load source video.',
-            })
-            setOriginalVideoFile(sourceFile)
-          } catch {
-            // Non-critical. originalVideoFile stays as the working video.
           }
+
+          setIsDraftLoading(false)
+          setScreen(SCREENS.EDITOR)
+        } catch {
+          setDraftLoadError('Could not restore draft video. The file may have expired.')
+          setPendingVideoDraftPayload(normalizedPayload)
+          setActiveDraftId(null)
+          setActiveDraftTitle(null)
+        } finally {
+          setIsDraftLoading(false)
         }
 
-        setIsDraftLoading(false)
-        setScreen(SCREENS.EDITOR)
-      } catch {
-        setDraftLoadError('Could not restore draft video. The file may have expired.')
-        setPendingVideoDraftPayload(payload)
-        setActiveDraftId(null)
-        setActiveDraftTitle(null)
-      } finally {
-        setIsDraftLoading(false)
+        return
       }
-
-      return
-    }
-  }, [gifSession, handleVideoSelect, resetImageEditingSessionState, resetSelection, restoreImageSession])
+    },
+    [gifSession, handleVideoSelect, resetImageEditingSessionState, resetSelection, restoreImageSession],
+  )
 
   useEffect(() => {
     onSelectCreation?.(handleLoadDraft)
@@ -404,246 +383,6 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
 
     return () => window.clearTimeout(timer)
   }, [isDraftLoading])
-
-  const handleSaveForLaterImage = useCallback(async () => {
-    if (!selectedMedia) return
-
-    setSaveForLaterError(null)
-    setSaveForLaterMessage(null)
-    setIsSavingDraft(true)
-
-    try {
-      const title = activeDraftTitle || defaultCreationTitle(selectedMedia)
-      const editorPayload = buildImageCreationPayload({
-        sourceMediaId: effectiveImageDraftSourceMediaId || effectiveBackendMediaId,
-        workingMediaId: effectiveBackendMediaId || effectiveImageDraftSourceMediaId,
-        preEditWorkingMediaId: editBaseMediaId || null,
-        preTextWorkingMediaId: preTextWorkingMediaId || null,
-        textOverlay: appliedTextOverlay || null,
-        lastCropBoxPx,
-        colorAdjustments,
-        selectedImageFilterPreset,
-        selectedPreset,
-        letterboxColor,
-      })
-
-      const body = { title, editorPayload, status: 'draft' }
-      const result = activeDraftId
-        ? await updateCreation(activeDraftId, body)
-        : await createCreation({ ...body, ownerKey: getOrCreateOwnerKey() })
-
-      const id = result?._id ?? result?.id
-      if (id) {
-        setActiveDraftId(String(id))
-      }
-      setActiveDraftTitle(title)
-
-      setSaveForLaterMessage('Draft saved. View it in My Creations.')
-      onDraftSaved?.()
-    } catch (err) {
-      setSaveForLaterError(err?.message || 'Could not save draft.')
-    } finally {
-      setIsSavingDraft(false)
-    }
-  }, [
-    activeDraftId,
-    activeDraftTitle,
-    appliedTextOverlay,
-    colorAdjustments,
-    editBaseMediaId,
-    effectiveBackendMediaId,
-    effectiveImageDraftSourceMediaId,
-    lastCropBoxPx,
-    letterboxColor,
-    onDraftSaved,
-    preTextWorkingMediaId,
-    selectedImageFilterPreset,
-    selectedMedia,
-    selectedPreset,
-  ])
-
-  const handleSaveForLaterVideo = useCallback(async () => {
-    if (!selectedMedia) return
-
-    setSaveForLaterError(null)
-    setSaveForLaterMessage(null)
-    setIsSavingDraft(true)
-
-    const sourceMediaId = effectiveVideoDraftSourceMediaId || null
-    const workingMediaId = backendVideoResult?.id || sourceMediaId
-
-    if (!workingMediaId) {
-      setSaveForLaterError('Video is not uploaded to the backend yet. Please wait and try saving again.')
-      setIsSavingDraft(false)
-      return
-    }
-
-    try {
-      const title = activeDraftTitle || defaultCreationTitle(selectedMedia)
-      const editorPayload = buildVideoCreationPayload({
-        sourceMediaId,
-        workingMediaId,
-        trimRange: gifSession.trimRange,
-        resizePreset: gifSession.resizePreset,
-        resizeBorderColor: gifSession.resizeBorderColor,
-        selectedSpeedPlaybackRate: gifSession.selectedSpeedPlaybackRate,
-        textOverlaySettings: gifSession.textOverlaySettings,
-        selectedFilterPreset: gifSession.selectedFilterPreset,
-      })
-
-      const body = { title, editorPayload, status: 'draft' }
-      const result = activeDraftId
-        ? await updateCreation(activeDraftId, body)
-        : await createCreation({ ...body, ownerKey: getOrCreateOwnerKey() })
-
-      const id = result?._id ?? result?.id
-      if (id) {
-        setActiveDraftId(String(id))
-      }
-      setActiveDraftTitle(title)
-
-      setSaveForLaterMessage('Draft saved. View it in My Creations.')
-      onDraftSaved?.()
-    } catch (err) {
-      setSaveForLaterError(err?.message || 'Could not save draft.')
-    } finally {
-      setIsSavingDraft(false)
-    }
-  }, [
-    activeDraftId,
-    activeDraftTitle,
-    backendVideoResult?.id,
-    effectiveVideoDraftSourceMediaId,
-    gifSession.resizeBorderColor,
-    gifSession.resizePreset,
-    gifSession.selectedFilterPreset,
-    gifSession.selectedSpeedPlaybackRate,
-    gifSession.textOverlaySettings,
-    gifSession.trimRange,
-    onDraftSaved,
-    selectedMedia,
-  ])
-
-  const handleOpenFilters = () => {
-    setScreen(SCREENS.FILTERS_MAIN)
-  }
-
-  const handleOpenSizes = () => {
-    setScreen(SCREENS.PRESET_SIZES)
-  }
-
-  const handleVideoPresetApply = useCallback(
-    async (result) => {
-      if (!result?.url) {
-        if (originalVideoFile) {
-          await handleVideoSelect(originalVideoFile, {
-            preserveSelectedSpeed: true,
-            preserveOriginalVideo: true,
-          })
-        }
-
-        gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.EDITOR)
-        setScreen(SCREENS.EDITOR)
-        return
-      }
-
-      const { file: filteredFile } = await convertBackendVideoResultToLocalMedia(result, {
-        fallbackFileName: `filtered-${result?.preset || 'video'}.mp4`,
-        fallbackMimeType: 'video/mp4',
-        fetchErrorMessage: 'Failed to load filtered video.',
-      })
-
-      const selection = await handleVideoSelect(filteredFile, {
-        preserveSelectedSpeed: true,
-        preserveOriginalVideo: true,
-        preserveDraftIdentity: true,
-      })
-
-      if (selection?.code !== MEDIA_SELECTION_CODES.OK) {
-        throw new Error('Filtered video could not be loaded in editor.')
-      }
-
-      gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.EDITOR)
-      setScreen(SCREENS.EDITOR)
-    },
-    [handleVideoSelect, gifSession, originalVideoFile],
-  )
-
-  const handleGifTrimApply = useCallback((nextRange) => {
-    gifSession.applyTrimRange(nextRange)
-  }, [gifSession])
-
-  const handleGifResizeApply = useCallback((nextResizeSettings) => {
-    gifSession.applyResizeSettings(nextResizeSettings)
-  }, [gifSession])
-
-  const handleCropApplyWrapped = useCallback(async (cropRequest) => {
-    const result = await handleCropApply(cropRequest)
-    setAppliedTextOverlay(null)
-    setPreTextWorkingMediaId(null)
-    return result
-  }, [handleCropApply])
-
-  const handlePresetSizeApply = async ({ preset, letterboxColor: nextLetterboxColor }) => {
-    const applied = await handleSizeSelect(preset, {
-      letterboxColor: nextLetterboxColor,
-    })
-
-    if (!applied) {
-      return
-    }
-
-    setLetterboxColor(nextLetterboxColor)
-    setScreen(SCREENS.EDITOR)
-  }
-
-  const handleAddTextScreenApply = async (textRequest) => {
-    const baseForText = preTextWorkingMediaId || effectiveBackendMediaId
-    if (!preTextWorkingMediaId && effectiveBackendMediaId) {
-      setPreTextWorkingMediaId(effectiveBackendMediaId)
-    }
-
-    const applied = await handleAddTextApply({
-      ...textRequest,
-      _overrideBaseMediaId: baseForText || undefined,
-    })
-
-    if (applied) {
-      setAppliedTextOverlay({
-        text: textRequest?.text ?? '',
-        font: textRequest?.font || textRequest?.fontFamily || 'Arial',
-        color: textRequest?.color || '#111111',
-        fontSize: textRequest?.fontSize ?? null,
-        x: textRequest?.x ?? 0.5,
-        y: textRequest?.y ?? 0.5,
-      })
-      setScreen(SCREENS.EDITOR)
-    }
-  }
-
-  const renderImageEditor = () => (
-    <ImageEditor
-      imageSrc={effectiveImageSrc}
-      cropSourceImageSrc={effectiveImageSrc}
-      initialCropPx={lastCropBoxPx}
-      onCropApply={handleCropApplyWrapped}
-      onResetCrop={resetCropToOriginal}
-      isUploading={isUploading}
-      isExporting={isExporting}
-      isResettingCrop={isResettingCrop}
-      uploadError={uploadError}
-      exportError={exportError}
-      onBack={handleBackToUpload}
-      onOpenFilters={handleOpenFilters}
-      onSize={handleOpenSizes}
-      onExport={handleExport}
-      onSaveForLater={handleSaveForLaterImage}
-      isSavingDraft={isSavingDraft}
-      saveDraftError={saveForLaterError}
-      saveDraftMessage={saveForLaterMessage}
-      showResetCrop={hasCropHistory}
-    />
-  )
 
   const renderCameraCapture = () => (
     <CameraCapture
@@ -695,138 +434,15 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
   )
 
   const renderNoSelectionFlow = () => {
-    if (screen === SCREENS.CAMERA) {
-      return renderCameraCapture()
-    }
-
-    if (screen === SCREENS.CAMERA_PREVIEW) {
-      return renderCameraPreview()
-    }
+    if (screen === SCREENS.CAMERA) return renderCameraCapture()
+    if (screen === SCREENS.CAMERA_PREVIEW) return renderCameraPreview()
 
     return (
       <>
-        {draftLoadError && (
-          <EditorStatus tone="error" spaced>{draftLoadError}</EditorStatus>
-        )}
+        {draftLoadError && <EditorStatus tone="error" spaced>{draftLoadError}</EditorStatus>}
         {renderMediaEntry()}
       </>
     )
-  }
-
-  const renderVideoFlow = () => {
-    const videoKey = selectedMedia
-      ? `${selectedMedia.name ?? selectedMedia.id ?? 'video'}-${selectedMedia.lastModified ?? ''}-${selectedMedia.size ?? ''}`
-      : 'video'
-
-    return (
-      <GifEditor
-        key={videoKey}
-        videoFile={selectedMedia}
-        gifSessionState={{
-          trimRange: gifSession.trimRange,
-          resizePreset: gifSession.resizePreset,
-          resizeBorderColor: gifSession.resizeBorderColor,
-          selectedSpeedPlaybackRate: gifSession.selectedSpeedPlaybackRate,
-          textOverlaySettings: gifSession.textOverlaySettings,
-        }}
-        onCancel={handleBackToUpload}
-        onCreateGif={(videoFile, trimOverrides) => gifSession.createAndExportGif(videoFile, trimOverrides)}
-        onOpenTrim={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.TRIM)}
-        onOpenResize={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.RESIZE)}
-        onOpenFilters={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.FILTERS_MAIN)}
-        onSaveForLater={handleSaveForLaterVideo}
-        isSavingDraft={isSavingDraft}
-        saveDraftError={saveForLaterError}
-        saveDraftMessage={saveForLaterMessage}
-      />
-    )
-  }
-
-  const renderImageFlow = () => {
-    switch (screen) {
-      case SCREENS.EDITOR:
-        return renderImageEditor()
-      case SCREENS.FILTERS_MAIN:
-        return (
-          <FilterMain
-            onPresetFilters={() => setScreen(SCREENS.IMAGE_PRESET_FILTERS)}
-            onText={() => setScreen(SCREENS.ADD_TEXT)}
-            onColorFilters={() => setScreen(SCREENS.COLOR_FILTERS)}
-            onCancel={() => setScreen(SCREENS.EDITOR)}
-          />
-        )
-      case SCREENS.IMAGE_PRESET_FILTERS:
-        return (
-          <PresetFilters
-            imageSrc={effectiveImageSrc}
-            selectedStyle={selectedImageFilterPreset}
-            previewSrc={presetFilterPreviewSrc}
-            onSelectStyle={selectImageFilterPreset}
-            onApply={async () => {
-              const applied = await applyImagePresetFilter()
-              if (applied) setScreen(SCREENS.EDITOR)
-            }}
-            onCancel={() => setScreen(SCREENS.FILTERS_MAIN)}
-            applyError={exportError}
-            previewError={presetFilterError}
-            isLoadingPreview={isLoadingPresetFilterPreview}
-          />
-        )
-      case SCREENS.ADD_TEXT: {
-        const preTextImageSrc = preTextWorkingMediaId
-          ? `${getBackendBaseUrl()}/api/media/${encodeURIComponent(preTextWorkingMediaId)}`
-          : effectiveImageSrc
-
-        const initialUiFontSize = appliedTextOverlay?.fontSize
-          ? Math.max(2, Math.round(appliedTextOverlay.fontSize / 5))
-          : undefined
-
-        return (
-          <AddText
-            imageSrc={preTextImageSrc}
-            initialText={appliedTextOverlay?.text ?? undefined}
-            initialFont={appliedTextOverlay?.font ?? undefined}
-            initialTextColor={appliedTextOverlay?.color ?? undefined}
-            initialFontSize={initialUiFontSize}
-            initialPlacement={appliedTextOverlay ? { x: appliedTextOverlay.x, y: appliedTextOverlay.y } : undefined}
-            onApply={handleAddTextScreenApply}
-            onBack={() => setScreen(SCREENS.FILTERS_MAIN)}
-            onCancel={() => setScreen(SCREENS.EDITOR)}
-            applyError={exportError}
-          />
-        )
-      }
-      case SCREENS.COLOR_FILTERS:
-        return (
-          <ColorFilters
-            imageSrc={effectiveImageSrc}
-            adjustments={colorAdjustments}
-            previewSrc={colorFilterPreviewSrc}
-            onAdjustmentsChange={updateColorAdjustments}
-            onApply={async () => {
-              const applied = await applyColorAdjustments()
-              if (applied) setScreen(SCREENS.EDITOR)
-            }}
-            onCancel={() => setScreen(SCREENS.EDITOR)}
-            applyError={exportError}
-            previewError={colorFilterError}
-            isLoadingPreview={isLoadingColorFilterPreview}
-          />
-        )
-      case SCREENS.PRESET_SIZES:
-        return (
-          <PresetSizes
-            imageSrc={effectiveImageSrc}
-            initialPreset={selectedPreset}
-            initialLetterboxColor={letterboxColor}
-            onApply={handlePresetSizeApply}
-            onCancel={() => setScreen(SCREENS.EDITOR)}
-            isBusy={isExporting}
-          />
-        )
-      default:
-        return renderImageEditor()
-    }
   }
 
   const renderContent = () => {
@@ -835,97 +451,43 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
     if (!selectedMedia) {
       content = renderNoSelectionFlow()
     } else if (mediaType === 'video') {
-      if (gifSession.activeTool === gifSession.GIF_FLOW_TOOLS.FILTERS_MAIN) {
-        content = (
-          <GifFilterMain
-            onPresetFilters={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.PRESET_FILTERS)}
-            onTextOverlay={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.TEXT)}
-            onSpeed={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.SPEED)}
-            onCancel={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.EDITOR)}
-          />
-        )
-      } else if (gifSession.activeTool === gifSession.GIF_FLOW_TOOLS.PRESET_FILTERS) {
-        const sourceVideoForFilters = originalVideoFile || selectedMedia
-
-        content = (
-          <VideoPresetFilters
-            videoFile={sourceVideoForFilters}
-            selectedFilter={gifSession.selectedFilterPreset}
-            onSelectFilter={(presetId) => gifSession.selectFilterPreset(sourceVideoForFilters, presetId)}
-            onApply={async () => {
-              if (gifSession.selectedFilterPreset === 'default') {
-                await gifSession.selectFilterPreset(sourceVideoForFilters, 'default')
-                await handleVideoPresetApply(null)
-                return
-              }
-
-              const preset = gifSession.selectedFilterPreset
-              const existingResult = gifSession.filterPreviewResult
-
-              try {
-                const result =
-                  existingResult?.preset === preset && existingResult?.url
-                    ? existingResult
-                    : await gifSession.applyVideoFilterAndReturn(sourceVideoForFilters, preset, existingResult)
-
-                await handleVideoPresetApply(result)
-              } catch {
-                // Errors are already surfaced elsewhere.
-              }
-            }}
-            onCancel={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.FILTERS_MAIN)}
-            isLoadingPreview={gifSession.isLoadingFilterPreview}
-            previewError={gifSession.filterPreviewError}
-            previewUrl={gifSession.filterPreviewUrl}
-          />
-        )
-      } else if (gifSession.activeTool === gifSession.GIF_FLOW_TOOLS.TEXT) {
-        content = (
-          <GifTextOverlayEditor
-            videoFile={selectedMedia}
-            initialSettings={gifSession.textOverlaySettings}
-            onChange={gifSession.updateGifTextOverlaySettings}
-            onApply={gifSession.applyGifTextOverlaySettings}
-            onBack={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.FILTERS_MAIN)}
-            onCancel={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.EDITOR)}
-          />
-        )
-      } else if (gifSession.activeTool === gifSession.GIF_FLOW_TOOLS.SPEED) {
-        content = (
-          <GifSpeedControls
-            videoFile={selectedMedia}
-            selectedSpeedPlaybackRate={gifSession.selectedSpeedPlaybackRate}
-            onSelectSpeed={gifSession.selectSpeed}
-            onApplySpeed={gifSession.applySpeed}
-            onBack={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.FILTERS_MAIN)}
-            onCancel={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.EDITOR)}
-          />
-        )
-      } else if (gifSession.activeTool === gifSession.GIF_FLOW_TOOLS.RESIZE) {
-        content = (
-          <GifResizePresets
-            initialPreset={gifSession.resizePreset}
-            initialBorderColor={gifSession.resizeBorderColor}
-            videoFile={selectedMedia}
-            onApply={handleGifResizeApply}
-            onCancel={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.EDITOR)}
-          />
-        )
-      } else if (gifSession.activeTool === gifSession.GIF_FLOW_TOOLS.TRIM) {
-        content = (
-          <GifTrimEditor
-            videoFile={selectedMedia}
-            initialTrimStart={gifSession.trimRange.start}
-            initialTrimEnd={gifSession.trimRange.end}
-            onApply={handleGifTrimApply}
-            onCancel={() => gifSession.openGifTool(gifSession.GIF_FLOW_TOOLS.EDITOR)}
-          />
-        )
-      } else {
-        content = renderVideoFlow()
-      }
+      content = (
+        <GifEditorFlow
+          gifSession={gifSession}
+          media={{ selectedMedia, backendVideoResult }}
+          originalVideoFile={originalVideoFile}
+          draft={{
+            activeDraftId,
+            activeDraftTitle,
+            effectiveVideoDraftSourceMediaId,
+            onActiveDraftSaved: handleActiveDraftSaved,
+          }}
+          onVideoSelect={handleVideoSelect}
+          onBack={handleBackToUpload}
+          onDraftSaved={onDraftSaved}
+        />
+      )
     } else {
-      content = renderImageFlow()
+      content = (
+        <ImageEditorFlow
+          imageSession={imageSession}
+          media={{ selectedMedia, isUploading, uploadError }}
+          draft={{
+            activeDraftId,
+            activeDraftTitle,
+            effectiveImageDraftSourceMediaId,
+            onActiveDraftSaved: handleActiveDraftSaved,
+          }}
+          textOverlay={{
+            appliedTextOverlay,
+            preTextWorkingMediaId,
+            onAppliedTextOverlayChange: setAppliedTextOverlay,
+            onPreTextWorkingMediaIdChange: setPreTextWorkingMediaId,
+          }}
+          onBack={handleBackToUpload}
+          onDraftSaved={onDraftSaved}
+        />
+      )
     }
 
     return (
@@ -936,9 +498,7 @@ function EditorContainer({ onDraftSaved, onSelectCreation }) {
           <div className="draft-restore-overlay">
             <div className="draft-restore-banner">
               <span className="draft-restore-banner__spinner" aria-hidden="true" />
-              <p className="draft-restore-banner__text">
-                Opening draft…
-              </p>
+              <p className="draft-restore-banner__text">Opening draft…</p>
               {shouldShowDraftLoadCancel && (
                 <button
                   type="button"
